refactor(testing): extract mock restore exports
