+++ src/testing/module-mock-lifecycle.ts
@@ -1,6 +1,7 @@
 import { createRequire } from "node:module"
 import { fileURLToPath } from "node:url"
 import { defaultGetCallerStack, resolveCallerUrlFromStack } from "./module-mock-stack"
+import { createRestoreExports } from "./module-mock-restore-exports"
 
 type MockModuleFactory = () => Record<string, unknown>
 
@@ -38,34 +39,6 @@ let originalLoadNonce = 0
 
 function toError(error: unknown): Error {
   return error instanceof Error ? error : new Error(String(error))
-}
-
-function isModuleExports(moduleValue: unknown): moduleValue is Record<string, unknown> {
-  return moduleValue !== null && typeof moduleValue === "object"
-}
-
-function isModuleNamespaceObject(moduleValue: Record<string, unknown>): boolean {
-  return Object.prototype.toString.call(moduleValue) === "[object Module]"
-}
-
-function createRestoreExports(moduleValue: unknown): Record<string, unknown> {
-  if (typeof moduleValue === "function") {
-    const functionExports = Object.assign({}, moduleValue)
-    return {
-      ...functionExports,
-      default: moduleValue,
-    }
-  }
-
-  if (isModuleExports(moduleValue)) {
-    if (isModuleNamespaceObject(moduleValue)) {
-      return { ...moduleValue }
-    }
-
-    return moduleValue
-  }
-
-  return { default: moduleValue }
 }
 
 function resolveWithBun(specifier: string, callerUrl: string): string {
